feat: duration-based load mode (-d) with graceful deadline drain
- -d/--duration runs a wall-clock window (30s, 5m, 1h30m); mutually
  exclusive with -n, minimum window 1s, both validated with actionable
  errors; -n default removed so 'no workload defined' is detectable
- Producer-only deadline check: workers, collector and CI gates are
  untouched; in-flight requests drain naturally at the deadline — no
  synthetic errors pollute stats
- Time-based progress bar (elapsed/total + live req/s) for duration
  mode; count-mode bar unchanged
- requested_duration_ms added to JSON report; quiet CSV appends a
  requested-seconds column only in duration mode
- Job.ID / Result.JobID widened to int64 (unbounded request counts on
  32-bit platforms during long soaks)
- Fixed latent run/run-dist shared flag variable: pflag writes defaults
  into the bound var at registration, so whichever init() ran last won
- Docs: CHANGELOG entries incl. breaking-change notes, README examples

+++ pkg/storm/bench_test.go
@@ -14,7 +14,7 @@ func makeResults(n int) []Result {
 	results := make([]Result, n)
 	for i := range results {
 		results[i] = Result{
-			JobID:      i + 1,
+			JobID:      int64(i + 1),
 			Method:     "GET",
 			StatusCode: 200,
 			Duration:   time.Duration(i%1000+1) * time.Millisecond,
